// apps/mobile/yukle-mobile/src/types/settlement.ts
export interface SettlementPreview {
  bidAmount: number;
  driverCommission: number;
  customerCommission: number;
  withholding: number;
  driverNet: number;
  customerTotal: number;
  platformRevenue: number;
  driverCommissionRate: number;
  customerCommissionRate: number;
  stopajRate: number;
}